0.10.0

+++ commands/admin/clear.js
@@ -1,25 +1,33 @@
-const { SlashCommandBuilder } = require("@discordjs/builders");
-const { options } = require("nodemon/lib/config");
+const {
+  SlashCommandBuilder
+} = require("@discordjs/builders");
 module.exports = {
-  permissions: "ADMINISTRATOR",
   enable: true,
   cooldown: 0,
   data: new SlashCommandBuilder()
     .setName("clear")
     .setDescription("Supprime le nombre de message demandé")
+    .setDefaultMemberPermissions(8)
     .addIntegerOption((option) =>
       option
-        .setName("number")
-        .setDescription("Nombre de message à supprimer")
-        .setRequired(true)
+      .setName("number")
+      .setDescription("Nombre de message à supprimer")
+      .setRequired(true)
     ),
   async execute(bot, interaction, database) {
     const amount = interaction.options.getInteger("number");
-    interaction.channel.bulkDelete(amount, true).then((messages) =>
+    if (amount > 100)
       interaction.reply({
-        content: `${messages.size} messages supprimés`,
-        ephemeral: true,
+        content: 'Impossible de supprimer autant de messages !'
       })
-    );
+    else {
+      interaction.channel.bulkDelete(amount, true).then((messages) =>
+        interaction.reply({
+          content: `${messages.size} messages supprimés`,
+          ephemeral: true,
+        })
+      );
+    }
+
   },
 };
